Bugfix: the aperture extension must use integers

+++ astrocut/cube_cut.py
@@ -288,7 +288,7 @@ class CutoutFactory():
         uncert_cutout = cutout[:, :, :, 1].transpose((2, 0, 1))
     
         # Making the aperture array
-        aperture = np.ones((xmax-xmin, ymax-ymin))
+        aperture = np.ones((xmax-xmin, ymax-ymin), dtype=np.int32)
 
         # Adding padding to the cutouts so that it's the expected size
         if padding.any():  # only do if we need to pad
